refactor(machine-a-sous): reorganize NewGameService import path and restore FirebaseListenerService
