+++ frontend/src/app/services/machine-a-sous/new-game.service.ts
